add crawler project to aws lambda function
