+++ package.json
@@ -18,6 +18,7 @@
   "homepage": "http://remoto.cc",
   "dependencies": {
     "async": "^2.6.0",
+    "aws-sdk": "^2.176.0",
     "crawler": "^1.1.2"
   }
 }
